+++ blochwave/changelog.md
@@ -1,4 +1,6 @@
 ### 1.0
+### Fri 12 Jan 16:40:48 GMT 2024
+- get_lattice is not loaded any more but computed with ut.get_lattice to fix bug when updating Nmax
 ### Thu  7 Dec 08:17:10 GMT 2023
 - added frame parameter and custom excitation error calculation
 ### Thu 30 Nov 15:21:35 GMT 2023